[hotfix][tests] Refactor MapR FS Tests
